+++ backend/src/main/java/moheng/applicationrunner/data/dev/KeywordDevApplicationRunner.java
@@ -1,11 +1,11 @@
-package moheng.applicationrunner.local;
+package moheng.applicationrunner.data.dev;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import moheng.applicationrunner.dto.KeywordRunner;
+import moheng.applicationrunner.data.dto.KeywordRunner;
 import moheng.keyword.domain.Keyword;
-import moheng.keyword.domain.TripKeyword;
 import moheng.keyword.domain.repository.KeywordRepository;
+import moheng.keyword.domain.TripKeyword;
 import moheng.keyword.domain.repository.TripKeywordRepository;
 import moheng.trip.domain.Trip;
 import moheng.trip.domain.repository.TripRepository;
@@ -20,17 +20,17 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Profile("local")
+@Profile("dev")
 @Order(3)
 @Component
-public class KeywordLocalApplicationRunner implements ApplicationRunner {
+public class KeywordDevApplicationRunner implements ApplicationRunner {
     private final KeywordRepository keywordRepository;
     private final TripRepository tripRepository;
     private final TripKeywordRepository tripKeywordRepository;
 
-    public KeywordLocalApplicationRunner(final KeywordRepository keywordRepository,
-                                        final TripRepository tripRepository,
-                                        final TripKeywordRepository tripKeywordRepository) {
+    public KeywordDevApplicationRunner(final KeywordRepository keywordRepository,
+                                       final TripRepository tripRepository,
+                                       final TripKeywordRepository tripKeywordRepository) {
         this.keywordRepository = keywordRepository;
         this.tripRepository = tripRepository;
         this.tripKeywordRepository = tripKeywordRepository;
