+++ backend/src/main/java/moheng/applicationrunner/data/dto/LiveInformationRunner.java
@@ -1,4 +1,4 @@
-package moheng.applicationrunner.dto;
+package moheng.applicationrunner.data.dto;
 
 import java.util.List;
