+++ backend/src/main/java/moheng/applicationrunner/data/prod/KeywordProdApplicationRunner.java
@@ -1,8 +1,8 @@
-package moheng.applicationrunner.dev;
+package moheng.applicationrunner.data.prod;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import moheng.applicationrunner.dto.KeywordRunner;
+import moheng.applicationrunner.data.dto.KeywordRunner;
 import moheng.keyword.domain.Keyword;
 import moheng.keyword.domain.repository.KeywordRepository;
 import moheng.keyword.domain.TripKeyword;
@@ -20,17 +20,17 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Profile("dev")
+@Profile("prod")
 @Order(3)
 @Component
-public class KeywordDevApplicationRunner implements ApplicationRunner {
+public class KeywordProdApplicationRunner implements ApplicationRunner {
     private final KeywordRepository keywordRepository;
     private final TripRepository tripRepository;
     private final TripKeywordRepository tripKeywordRepository;
 
-    public KeywordDevApplicationRunner(final KeywordRepository keywordRepository,
-                                       final TripRepository tripRepository,
-                                       final TripKeywordRepository tripKeywordRepository) {
+    public KeywordProdApplicationRunner(final KeywordRepository keywordRepository,
+                                        final TripRepository tripRepository,
+                                        final TripKeywordRepository tripKeywordRepository) {
         this.keywordRepository = keywordRepository;
         this.tripRepository = tripRepository;
         this.tripKeywordRepository = tripKeywordRepository;
